Increase Docker machine to medium+ with 6GB RAM
We use a Docker medium machine with 4GB and this might cause an OOM when we allow to use 100% of the memory.
I've increased the Docker machine to medium+ with 6GB RAM to have some space with 4GB used by Node.
Split LUNR site generator installation into separate command for better debugging.