docs: Remove Zenith and Lokahi/Cloud from docs
The documentation for Zenith is empty right now. It was added to make sure the documentation
workflow is working as expected. We can add it again as soon we have documentation we can publish.

I have also removed Lokahi/Cloud from the docs, because we have removed all traces from Lokahi/Cloud and
links in the docs are going to the GitHub wiki page which ends up in 404 not found pages.

Signed-off-by: Ronny Trommer <ronny@no42.org>